Add initial part of submission box redesign
By default nothing is selected, so allows partial submissions of only a title, or only a thumbnail.

Switches selected visual to be a border instead of a size change.

Changes current time thumbnail to add new boxes to the UI.

Fix title change listener being out of date in submission component